requirements-gcp.txt
# Google Cloud Platform optimized requirements for CloudFace AI
# Lighter versions for better deployment performance

# Core dependencies
flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0

# Image processing (lighter versions)
Pillow==10.0.1
opencv-python-headless==4.8.1.78

# Face recognition (core - keep essential only)
face_recognition==1.3.0

# Vector databases (CPU optimized)
faiss-cpu==1.7.4

# Deep learning frameworks (lighter for testing)
torch==2.0.1+cpu --find-links https://download.pytorch.org/whl/torch_stable.html
torchvision==0.15.2+cpu --find-links https://download.pytorch.org/whl/torch_stable.html

# Face recognition models (essential only)
facenet-pytorch==2.5.3

# MediaPipe for face alignment
mediapipe==0.10.8

# Database and storage
firebase-admin==6.2.0

# Google Cloud specific
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-storage==2.10.0

# Payment gateways
razorpay==1.3.0

# Utilities
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.11.1

# Production server
waitress==2.1.2
